refactor: move search code to its own file
