+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod raw_types;
+pub mod search;
 pub mod types;
 pub mod util;
 pub mod wrapper;
