+++ src/wrapper.rs
@@ -14,6 +14,7 @@ use crate::raw_types::{
     RawCoursePrerequisite, RawDepartmentElement, RawEvent, RawPrerequisite, RawScheduledMeeting,
     RawSubjectElement, RawWebRegMeeting, RawWebRegSearchResultItem,
 };
+use crate::search::{DayOfWeek, SearchType};
 use crate::types::{
     CoursePrerequisite, CourseSection, EnrollmentStatus, Event, Meeting, MeetingDay,
     PrerequisiteInfo, ScheduledSection,
@@ -1173,9 +1174,10 @@ impl WebRegWrapper {
     /// ```rust,no_run
     /// use reqwest::Client;
     /// use webweg::wrapper::{SearchType, WebRegWrapper};
-    ///
+    ///use webweg::search::SearchType;
     /// # #[tokio::main(flavor = "current_thread")]
     /// # async fn main() {
+    ///
     /// let wrapper = WebRegWrapper::new(Client::new(), "my cookies".to_string(), "FA22");
     /// let res = wrapper.search_courses_detailed(SearchType::BySection("1234567")).await;
     ///
@@ -1191,6 +1193,7 @@ impl WebRegWrapper {
     /// for any sections with IDs `1234567` or `115123` or `2135`.
     /// ```rust,no_run
     /// use reqwest::Client;
+    /// use webweg::search::SearchType;
     /// use webweg::wrapper::{SearchType, WebRegWrapper};
     ///
     /// # #[tokio::main(flavor = "current_thread")]
@@ -1217,6 +1220,7 @@ impl WebRegWrapper {
     /// - ends no later than 5:30pm.
     /// ```rust,no_run
     /// use reqwest::Client;
+    /// use webweg::search::{CourseLevelFilter, DayOfWeek, SearchRequestBuilder, SearchType};
     /// use webweg::wrapper::{
     ///     CourseLevelFilter, DayOfWeek, SearchRequestBuilder, SearchType, WebRegWrapper
     /// };
@@ -2695,267 +2699,6 @@ pub struct EventAdd<'a> {
     /// The minute end time. For example, if the event ends at 3:50 PM,
     /// use `50`.
     pub end_min: i16,
-}
-
-/// Used to construct search requests for the `search_courses` function.
-#[derive(Clone)]
-pub struct SearchRequestBuilder {
-    subjects: Vec<String>,
-    courses: Vec<String>,
-    departments: Vec<String>,
-    instructor: Option<String>,
-    title: Option<String>,
-    level_filter: u32,
-    days: u32,
-    start_time: Option<(u32, u32)>,
-    end_time: Option<(u32, u32)>,
-    only_open: bool,
-}
-
-impl SearchRequestBuilder {
-    /// Creates a new instance of the `SearchRequestBuilder`, which is used to search for specific
-    /// courses.
-    ///
-    /// # Returns
-    /// The empty `SearchRequestBuilder`.
-    pub fn new() -> Self {
-        Self {
-            subjects: vec![],
-            courses: vec![],
-            departments: vec![],
-            instructor: None,
-            title: None,
-            level_filter: 0,
-            days: 0,
-            start_time: None,
-            end_time: None,
-            only_open: false,
-        }
-    }
-
-    /// Adds a subject to this search request. Valid search requests are uppercase and at most
-    /// 4 characters long. Some examples include `MATH` or `CSE`.
-    ///
-    /// # Parameters
-    /// - `subject`: The subject.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn add_subject(mut self, subject: impl Into<String>) -> Self {
-        let s: String = subject.into();
-        if s.len() > 4 {
-            return self;
-        }
-
-        self.subjects.push(s.to_uppercase());
-        self
-    }
-
-    /// Adds a course (either a subject code, course code, or both) to the search request. Some
-    /// examples include `20E`, `math 20d`, `101`, `CSE`.
-    ///
-    /// # Parameters
-    /// - `course`: The course.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn add_course(mut self, course: impl Into<String>) -> Self {
-        self.courses.push(course.into());
-        self
-    }
-
-    /// Adds a department to the search request. Valid search requests are uppercase and at most 4
-    /// characters long. Some examples include `MATH` or `CSE`.
-    ///
-    /// # Parameters
-    /// - `department`: The department.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn add_department(mut self, department: impl Into<String>) -> Self {
-        let d: String = department.into();
-        if d.len() > 4 {
-            return self;
-        }
-
-        self.departments.push(d.to_uppercase());
-        self
-    }
-
-    /// Sets the instructor to the specified instructor.
-    ///
-    /// # Parameters
-    /// - `instructor`: The instructor. This should be formatted in `Last Name, First Name` form.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn set_instructor(mut self, instructor: impl Into<String>) -> Self {
-        self.instructor = Some(instructor.into());
-        self
-    }
-
-    /// Sets the course title to the specified title. Some examples could be `differential equ`,
-    /// `data structures`, `algorithms`, and so on.
-    ///
-    /// # Parameters
-    /// - `title`: The title of the course.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn set_title(mut self, title: impl Into<String>) -> Self {
-        self.title = Some(title.into());
-        self
-    }
-
-    /// Restrict search results to to the specified filter. This can be applied multiple times.
-    ///
-    /// # Parameters
-    /// - `filter`: The filter.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn filter_courses_by(mut self, filter: CourseLevelFilter) -> Self {
-        self.level_filter |= match filter {
-            CourseLevelFilter::LowerDivision => 1 << 11,
-            CourseLevelFilter::FreshmenSeminar => 1 << 10,
-            CourseLevelFilter::LowerDivisionIndepStudy => 1 << 9,
-            CourseLevelFilter::UpperDivision => 1 << 8,
-            CourseLevelFilter::Apprenticeship => 1 << 7,
-            CourseLevelFilter::UpperDivisionIndepStudy => 1 << 6,
-            CourseLevelFilter::Graduate => 1 << 5,
-            CourseLevelFilter::GraduateIndepStudy => 1 << 4,
-            CourseLevelFilter::GraduateResearch => 1 << 3,
-            CourseLevelFilter::Lvl300 => 1 << 2,
-            CourseLevelFilter::Lvl400 => 1 << 1,
-            CourseLevelFilter::Lvl500 => 1 << 0,
-        };
-
-        self
-    }
-
-    /// Only shows courses based on the specified day(s).
-    ///
-    /// # Parameters
-    /// - `day`: The day.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn apply_day(mut self, day: DayOfWeek) -> Self {
-        let day = match day {
-            DayOfWeek::Monday => 1,
-            DayOfWeek::Tuesday => 2,
-            DayOfWeek::Wednesday => 3,
-            DayOfWeek::Thursday => 4,
-            DayOfWeek::Friday => 5,
-            DayOfWeek::Saturday => 6,
-            DayOfWeek::Sunday => 7,
-        };
-
-        self.days |= 1 << (7 - day);
-        self
-    }
-
-    /// Sets the start time to the specified time.
-    ///
-    /// # Parameters
-    /// - `hour`: The hour. This should be between 0 and 23, inclusive.
-    /// - `min`: The minute. This should be between 0 and 59, inclusive.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn set_start_time(mut self, hour: u32, min: u32) -> Self {
-        if hour > 23 || min > 59 {
-            return self;
-        }
-
-        self.start_time = Some((hour, min));
-        self
-    }
-
-    /// Sets the end time to the specified time.
-    ///
-    /// # Parameters
-    /// - `hour`: The hour. This should be between 0 and 23, inclusive.
-    /// - `min`: The minute. This should be between 0 and 59, inclusive.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn set_end_time(mut self, hour: u32, min: u32) -> Self {
-        if hour > 23 || min > 59 {
-            return self;
-        }
-
-        self.end_time = Some((hour, min));
-        self
-    }
-
-    /// Whether to only show sections with open seats.
-    ///
-    /// # Returns
-    /// The `SearchRequestBuilder`
-    pub fn only_allow_open(mut self) -> Self {
-        self.only_open = true;
-        self
-    }
-}
-
-impl Default for SearchRequestBuilder {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// The day of week enum, which designates what days you want
-/// to filter specific sections by.
-pub enum DayOfWeek {
-    Monday,
-    Tuesday,
-    Wednesday,
-    Thursday,
-    Friday,
-    Saturday,
-    Sunday,
-}
-
-/// The course level filter enum, which can be used to filter
-/// specific sections by.
-pub enum CourseLevelFilter {
-    /// Level 1-99 courses.
-    LowerDivision,
-    /// Level 87, 90 courses.
-    FreshmenSeminar,
-    /// Level 99 courses.
-    LowerDivisionIndepStudy,
-    /// Level 100-198 courses
-    UpperDivision,
-    /// Level 195 courses
-    Apprenticeship,
-    /// Level 199 courses
-    UpperDivisionIndepStudy,
-    /// Level 200-297 courses
-    Graduate,
-    /// Level 298 courses
-    GraduateIndepStudy,
-    /// Level 299 courses
-    GraduateResearch,
-    /// Level 300+ courses
-    Lvl300,
-    /// Level 400+ courses
-    Lvl400,
-    /// Level 500+ courses
-    Lvl500,
-}
-
-/// Lets you choose how you want to search for a course.
-pub enum SearchType<'a> {
-    /// Searches for a course by section ID.
-    BySection(&'a str),
-
-    /// Searches for a course by more than one section ID.
-    ByMultipleSections(&'a [&'a str]),
-
-    /// Searches for a (set of) course(s) by multiple specifications.
-    Advanced(&'a SearchRequestBuilder),
 }
 
 /// The possible grading options.
